+++ voe/__about__.py
@@ -11,7 +11,7 @@ __summary__ = (
 )
 __uri__ = 'https://github.com/ripiuk/voe_disconnect'
 
-__version__ = '1.0.0'
+__version__ = '1.1.0'
 
 __author__ = 'ripiuk'
 __email__ = ''