+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+### [1.1.0] (2024-06-23)
+
+* Changed daily run actions workflow schedule
+* Added extra VOE request headers
+* Fixed empty telegram message
+* Changed search params settings structure
+* Added multiple queue handling
+
+---
+
 ### [1.0.0] (2024-06-21)
 
 * Added project configuration
@@ -8,3 +18,4 @@
 ---
 
 [1.0.0]:https://github.com/ripiuk/voe_disconnect/pull/1
+[1.1.0]:https://github.com/ripiuk/voe_disconnect/pull/2